refractor: TemplateCreatorPage -> createUpdatedTemplate logic moved to dataHelpers page
